refactor tests to different package so they would actually run
